@@ -7,5 +7,5 @@ enum DynamicsType {
 
 extension BusinessTypeExtension on DynamicsType {
   String get values => ['all', 'video', 'pgc', 'article'][index];
-  String get labels => ['全部', '视频', '追番', '专栏'][index];
+  String get labels => ['全部', '投稿', '番剧', '专栏'][index];
 }
