@@ -17,6 +17,19 @@ List defaultNavigationBars = [
   {
     'id': 1,
     'icon': const Icon(
+      Icons.trending_up,
+      size: 21,
+    ),
+    'selectIcon': const Icon(
+      Icons.trending_up_outlined,
+      size: 21,
+    ),
+    'label': "排行榜",
+    'count': 0,
+  },
+  {
+    'id': 2,
+    'icon': const Icon(
       Icons.motion_photos_on_outlined,
       size: 21,
     ),
@@ -28,7 +41,7 @@ List defaultNavigationBars = [
     'count': 0,
   },
   {
-    'id': 2,
+    'id': 3,
     'icon': const Icon(
       Icons.video_collection_outlined,
       size: 20,
