@@ -16,7 +16,7 @@ class FavDetailController extends GetxController {
   RxMap favInfo = {}.obs;
   RxList favList = [].obs;
   RxString loadingText = '加载中...'.obs;
-  int mediaCount = 0;
+  RxInt mediaCount = 0.obs;
 
   @override
   void onInit() {
@@ -29,7 +29,7 @@ class FavDetailController extends GetxController {
   }
 
   Future<dynamic> queryUserFavFolderDetail({type = 'init'}) async {
-    if (type == 'onLoad' && favList.length >= mediaCount) {
+    if (type == 'onLoad' && favList.length >= mediaCount.value) {
       loadingText.value = '没有更多了';
       return;
     }
@@ -43,11 +43,11 @@ class FavDetailController extends GetxController {
       favInfo.value = res['data'].info;
       if (currentPage == 1 && type == 'init') {
         favList.value = res['data'].medias;
-        mediaCount = res['data'].info['media_count'];
+        mediaCount.value = res['data'].info['media_count'];
       } else if (type == 'onLoad') {
         favList.addAll(res['data'].medias);
       }
-      if (favList.length >= mediaCount) {
+      if (favList.length >= mediaCount.value) {
         loadingText.value = '没有更多了';
       }
     }
