@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:get/get.dart';
+import 'package:media_kit/media_kit.dart';
 import 'package:pilipala/common/widgets/network_img_layer.dart';
 import 'package:pilipala/models/user/fav_folder.dart';
 import 'package:pilipala/pages/main/index.dart';
@@ -102,7 +103,11 @@ class _MediaPageState extends State<MediaPage>
             ],
             Obx(() => mediaController.userLogin.value
                 ? favFolder(mediaController, context)
-                : const SizedBox())
+                : const SizedBox()),
+            SizedBox(
+              height: MediaQuery.of(context).padding.bottom +
+                  kBottomNavigationBarHeight,
+            )
           ],
         ),
       ),
