@@ -1,17 +1,20 @@
 import 'package:audio_video_progress_bar/audio_video_progress_bar.dart';
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:get/get.dart';
 import 'package:nil/nil.dart';
 import 'package:pilipala/plugin/pl_player/index.dart';
-import 'package:pilipala/plugin/pl_player/widgets/play_pause_btn.dart';
 import 'package:pilipala/utils/feed_back.dart';
 
 class BottomControl extends StatelessWidget implements PreferredSizeWidget {
   final PlPlayerController? controller;
   final Function? triggerFullScreen;
-  const BottomControl({this.controller, this.triggerFullScreen, Key? key})
-      : super(key: key);
+  final List<Widget>? buildBottomControl;
+  const BottomControl({
+    this.controller,
+    this.triggerFullScreen,
+    this.buildBottomControl,
+    Key? key,
+  }) : super(key: key);
 
   @override
   Size get preferredSize => const Size(double.infinity, kToolbarHeight);
@@ -20,11 +23,6 @@ class BottomControl extends StatelessWidget implements PreferredSizeWidget {
   Widget build(BuildContext context) {
     Color colorTheme = Theme.of(context).colorScheme.primary;
     final _ = controller!;
-    const textStyle = TextStyle(
-      color: Colors.white,
-      fontSize: 12,
-    );
-
     return Container(
       color: Colors.transparent,
       height: 90,
@@ -71,86 +69,89 @@ class BottomControl extends StatelessWidget implements PreferredSizeWidget {
             },
           ),
           Row(
-            children: [
-              PlayOrPauseButton(
-                controller: _,
-              ),
-              const SizedBox(width: 4),
-              // 播放时间
-              Obx(() {
-                return Text(
-                  _.durationSeconds.value >= 3600
-                      ? printDurationWithHours(
-                          Duration(seconds: _.positionSeconds.value))
-                      : printDuration(
-                          Duration(seconds: _.positionSeconds.value)),
-                  style: textStyle,
-                );
-              }),
-              const SizedBox(width: 2),
-              const Text('/', style: textStyle),
-              const SizedBox(width: 2),
-              Obx(
-                () => Text(
-                  _.durationSeconds.value >= 3600
-                      ? printDurationWithHours(
-                          Duration(seconds: _.durationSeconds.value))
-                      : printDuration(
-                          Duration(seconds: _.durationSeconds.value)),
-                  style: textStyle,
-                ),
-              ),
-              const Spacer(),
-              // 倍速
-              // Obx(
-              //   () => SizedBox(
-              //     width: 45,
-              //     height: 34,
-              //     child: TextButton(
-              //       style: ButtonStyle(
-              //         padding: MaterialStateProperty.all(EdgeInsets.zero),
-              //       ),
-              //       onPressed: () {
-              //         _.togglePlaybackSpeed();
-              //       },
-              //       child: Text(
-              //         '${_.playbackSpeed.toString()}X',
-              //         style: textStyle,
-              //       ),
-              //     ),
-              //   ),
-              // ),
-              SizedBox(
-                height: 30,
-                child: TextButton(
-                  onPressed: () => _.toggleVideoFit(),
-                  style: ButtonStyle(
-                    padding: MaterialStateProperty.all(EdgeInsets.zero),
-                  ),
-                  child: Obx(
-                    () => Text(
-                      _.videoFitDEsc.value,
-                      style: const TextStyle(color: Colors.white, fontSize: 13),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(width: 10),
-              // 全屏
-              Obx(
-                () => ComBtn(
-                  icon: Icon(
-                    _.isFullScreen.value
-                        ? FontAwesomeIcons.compress
-                        : FontAwesomeIcons.expand,
-                    size: 15,
-                    color: Colors.white,
-                  ),
-                  fuc: () => triggerFullScreen!(),
-                ),
-              ),
-            ],
+            children: [...buildBottomControl!],
           ),
+          // Row(
+          //   children: [
+          //     PlayOrPauseButton(
+          //       controller: _,
+          //     ),
+          //     const SizedBox(width: 4),
+          //     // 播放时间
+          //     Obx(() {
+          //       return Text(
+          //         _.durationSeconds.value >= 3600
+          //             ? printDurationWithHours(
+          //                 Duration(seconds: _.positionSeconds.value))
+          //             : printDuration(
+          //                 Duration(seconds: _.positionSeconds.value)),
+          //         style: textStyle,
+          //       );
+          //     }),
+          //     const SizedBox(width: 2),
+          //     const Text('/', style: textStyle),
+          //     const SizedBox(width: 2),
+          //     Obx(
+          //       () => Text(
+          //         _.durationSeconds.value >= 3600
+          //             ? printDurationWithHours(
+          //                 Duration(seconds: _.durationSeconds.value))
+          //             : printDuration(
+          //                 Duration(seconds: _.durationSeconds.value)),
+          //         style: textStyle,
+          //       ),
+          //     ),
+          //     const Spacer(),
+          //     // 倍速
+          //     // Obx(
+          //     //   () => SizedBox(
+          //     //     width: 45,
+          //     //     height: 34,
+          //     //     child: TextButton(
+          //     //       style: ButtonStyle(
+          //     //         padding: MaterialStateProperty.all(EdgeInsets.zero),
+          //     //       ),
+          //     //       onPressed: () {
+          //     //         _.togglePlaybackSpeed();
+          //     //       },
+          //     //       child: Text(
+          //     //         '${_.playbackSpeed.toString()}X',
+          //     //         style: textStyle,
+          //     //       ),
+          //     //     ),
+          //     //   ),
+          //     // ),
+          //     SizedBox(
+          //       height: 30,
+          //       child: TextButton(
+          //         onPressed: () => _.toggleVideoFit(),
+          //         style: ButtonStyle(
+          //           padding: MaterialStateProperty.all(EdgeInsets.zero),
+          //         ),
+          //         child: Obx(
+          //           () => Text(
+          //             _.videoFitDEsc.value,
+          //             style: const TextStyle(color: Colors.white, fontSize: 13),
+          //           ),
+          //         ),
+          //       ),
+          //     ),
+          //     const SizedBox(width: 10),
+          //     // 全屏
+          //     Obx(
+          //       () => ComBtn(
+          //         icon: Icon(
+          //           _.isFullScreen.value
+          //               ? FontAwesomeIcons.compress
+          //               : FontAwesomeIcons.expand,
+          //           size: 15,
+          //           color: Colors.white,
+          //         ),
+          //         fuc: () => triggerFullScreen!(),
+          //       ),
+          //     ),
+          //   ],
+          // ),
           const SizedBox(height: 12),
         ],
       ),
